+++ app/src/main/java/me/chamada/ft_hangouts/data/model/contact/Contact.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "contacts")
 data class Contact(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String = "",
     @ColumnInfo(name = "phone_number") val phoneNumber: String = ""
 )
