+++ app/src/main/java/me/chamada/ft_hangouts/ui/contacts/DetailsFragment.kt
@@ -19,7 +19,7 @@ import me.chamada.ft_hangouts.ui.DeleteDialogFragment
 
 class DetailsFragment : Fragment(), MenuProvider, DeleteDialogFragment.OnConfirmListener {
     private val viewModel: ContactViewModel by activityViewModels()
-    private var contactId: Int? = null
+    private var contactId: Long? = null
 
     private val deleteDialogFragment = DeleteDialogFragment(this)
